+++ architecture/repository/user/DeleteByID.go
@@ -1,5 +1,5 @@
 package user
 
-func (u *UserRepo) DeleteByID(id int) error {
+func (u *UserRepo) DeleteByID(id int64) error {
 	return nil
 }
