+++ architecture/service/user/DeleteByID.go
@@ -1,5 +1,5 @@
 package user
 
-func (u *UserService) DeleteByID(id int) error {
+func (u *UserService) DeleteByID(id int64) error {
 	return nil
 }
