+++ architecture/service/user/GetByID.go
@@ -2,6 +2,6 @@ package user
 
 import model "forum/architecture/models"
 
-func (u *UserService) GetByID(id int) (*model.User, error) {
+func (u *UserService) GetByID(id int64) (*model.User, error) {
 	return nil, nil
 }
